make platform dependent millis timer, as trait
