+++ src/mesh_lib/millis/mod.rs
@@ -1,6 +1,4 @@
-mod atmega328p;
+pub mod traits;
+pub mod types;
 
-#[allow(non_camel_case_types)]
-pub type ms = u32;
-
-pub use atmega328p::{millis, millis_init};
+pub use types::ms;
