+++ src/mesh_lib/millis/traits.rs
@@ -1,0 +1,7 @@
+use super::ms;
+
+pub trait PlatformTime {
+    type Timer;
+    fn init(timer: Self::Timer);
+    fn millis() -> ms;
+}
