+++ src/mesh_lib/millis/types.rs
@@ -1,0 +1,2 @@
+#[allow(non_camel_case_types)]
+pub type ms = u32;
